links_spec refactor pt 1
